+++ src/main/java/thelm/packagedavaritia/recipe/ExtremePackageRecipeType.java
@@ -3,7 +3,7 @@ package thelm.packagedavaritia.recipe;
 import java.util.List;
 import java.util.stream.IntStream;
 
-import avaritia.block.ModBlocks;
+import avaritia.init.ModTypes;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.ints.IntRBTreeSet;
@@ -87,7 +87,7 @@ public class ExtremePackageRecipeType implements IPackageRecipeType {
 
 	@Override
 	public Object getRepresentation() {
-		return new ItemStack(ModBlocks.extreme_crafting_table.get());
+		return new ItemStack(ModTypes.extreme_table.get());
 	}
 
 	@Override