+++ src/main/java/thelm/packagedavaritia/recipe/ExtremePackageRecipeType.java
@@ -33,6 +33,7 @@ public class ExtremePackageRecipeType implements IPackageRecipeType {
 	public static final ResourceLocation NAME = ResourceLocation.parse("packagedavaritia:extreme");
 	public static final IntSet SLOTS;
 	public static final List<ResourceLocation> CATEGORIES = List.of(ResourceLocation.parse("avaritia:extreme_craft"));
+	public static final List<ResourceLocation> CATEGORIES_EMI = List.of(ResourceLocation.parse("avaritia:extreme_crafting_table"));
 	public static final Vec3i COLOR = new Vec3i(139, 139, 139);
 	public static final Vec3i COLOR_DISABLED = new Vec3i(64, 64, 64);
 
@@ -86,6 +87,11 @@ public class ExtremePackageRecipeType implements IPackageRecipeType {
 	@Override
 	public List<ResourceLocation> getJEICategories() {
 		return CATEGORIES;
+	}
+
+	@Override
+	public List<ResourceLocation> getEMICategories() {
+		return CATEGORIES_EMI;
 	}
 
 	@Override