+++ src/main/java/thelm/packagedavaritia/recipe/NetherPackageRecipeType.java
@@ -32,6 +32,7 @@ public class NetherPackageRecipeType implements IPackageRecipeType {
 	public static final ResourceLocation NAME = ResourceLocation.parse("packagedavaritia:nether");
 	public static final IntSet SLOTS;
 	public static final List<ResourceLocation> CATEGORIES = List.of(ResourceLocation.parse("avaritia:nether_craft"));
+	public static final List<ResourceLocation> CATEGORIES_EMI = List.of(ResourceLocation.parse("avaritia:nether_crafting_table"));
 	public static final Vec3i COLOR = new Vec3i(139, 139, 139);
 	public static final Vec3i COLOR_DISABLED = new Vec3i(64, 64, 64);
 
@@ -89,6 +90,11 @@ public class NetherPackageRecipeType implements IPackageRecipeType {
 	@Override
 	public List<ResourceLocation> getJEICategories() {
 		return CATEGORIES;
+	}
+
+	@Override
+	public List<ResourceLocation> getEMICategories() {
+		return CATEGORIES_EMI;
 	}
 
 	@Override