+++ src/main/java/thelm/packagedavaritia/event/CommonEventHandler.java
@@ -12,6 +12,7 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.DeferredRegister;
 import thelm.packagedauto.util.ApiImpl;
@@ -32,6 +33,7 @@ import thelm.packagedavaritia.recipe.EndPackageRecipeType;
 import thelm.packagedavaritia.recipe.ExtremePackageRecipeType;
 import thelm.packagedavaritia.recipe.NetherPackageRecipeType;
 import thelm.packagedavaritia.recipe.SculkPackageRecipeType;
+import thelm.packagedavaritia.variant.VariantChecker;
 
 public class CommonEventHandler {
 
@@ -87,6 +89,11 @@ public class CommonEventHandler {
 					output.accept(ExtremeCrafterBlock.ITEM_INSTANCE);
 				}).
 				build());
+	}
+
+	@SubscribeEvent
+	public void onConstructMod(FMLConstructModEvent event) {
+		VariantChecker.check();
 	}
 
 	@SubscribeEvent