+++ src/main/java/thelm/packagedavaritia/recipe/ITablePackageRecipeInfo.java
@@ -2,7 +2,7 @@ package thelm.packagedavaritia.recipe;
 
 import java.util.List;
 
-import committee.nova.mods.avaritia.common.crafting.recipe.BaseTableCraftingRecipe;
+import committee.nova.mods.avaritia.common.crafting.recipe.ITierCraftingRecipe;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import thelm.packagedauto.api.IPackageRecipeInfo;
@@ -13,7 +13,7 @@ public interface ITablePackageRecipeInfo extends IPackageRecipeInfo {
 
 	ItemStack getOutput();
 
-	BaseTableCraftingRecipe getRecipe();
+	ITierCraftingRecipe getRecipe();
 
 	Container getMatrix();