+++ changelog.md
@@ -1,1 +1,1 @@
-Added support for the 5x5 and 7x7 tables  
+Adjust block strength values