+++ changelog.md
@@ -1,1 +1,2 @@
-1.19.2 Avaritia Universal  
+Adjust block strength values  
+Update Avaritia Universal