+++ changelog.md
@@ -1,1 +1,1 @@
-Updated to Re:Avaritia 1.3.9.x, and added new table recipe types  
+Adjust block strength values