+++ changelog.md
@@ -1,2 +1,1 @@
-Add Avaritia variant check  
-Adjust textures  
+Update Re:Avaritia to 1.3.9.1